Fixed bundle inclusion of ogc-schemas (#2093)
- Removed the inclusion of ogc-schemas in the default bundle
- Improved error handling and generalized OGC errors management (return 200 with ExceptionReport)
