+++ web/client/epics/wfsquery.js
@@ -16,7 +16,8 @@ const {paginationInfo, isDescribeLoaded, describeSelector} = require('../selecto
 const FilterUtils = require('../utils/FilterUtils');
 const assign = require('object-assign');
 
-const {isString, isObject} = require('lodash');
+const {isObject} = require('lodash');
+const {interceptOGCError} = require('../utils/ObservableUtils');
 // this is a workaround for https://osgeo-org.atlassian.net/browse/GEOS-7233. can be removed when fixed
 const workaroundGEOS7233 = ({totalFeatures, features, ...rest}, {startIndex, maxFeatures}, originalSize) => {
     if (originalSize > totalFeatures && originalSize === startIndex + features.length && totalFeatures === features.length) {
@@ -159,12 +160,6 @@ const getWFSFeature = (searchUrl, filterObj) => {
         }));
 };
 
-const getWFSResponseException = (response, code) => {
-    const {unmarshaller} = require('../utils/ogc/WFS');
-    const json = isString(response.data) ? unmarshaller.unmarshalString(response.data) : null;
-    return json && json.value && json.value.exception && json.value.exception[0] && json.value.exception[0].TYPE_NAME === 'OWS_1_0_0.ExceptionType' && json.value.exception[0].exceptionCode === code;
-};
-
 const getFirstAttribute = (state)=> {
     return state.query && state.query.featureTypes && state.query.featureTypes[state.query.typeName] && state.query.featureTypes[state.query.typeName].attributes && state.query.featureTypes[state.query.typeName].attributes[0] && state.query.featureTypes[state.query.typeName].attributes[0].attribute || null;
 };
@@ -178,11 +173,11 @@ const retryWithForcedSortOptions = (action, store) => {
     return getWFSFeature(action.searchUrl, assign(action.filterObj, {
         sortOptions
     }))
+        .let(interceptOGCError)
         .map((newResponse) => {
-            const newError = getWFSResponseException(newResponse, 'NoApplicableCode');
             const state = store.getState();
             const data = workaroundGEOS7233(newResponse.data, action.filterObj.pagination, paginationInfo.totalFeatures(state));
-            return !newError ? querySearchResponse(data, action.searchUrl, action.filterObj) : queryError('No sortable request');
+            return querySearchResponse(data, action.searchUrl, action.filterObj);
         })
         .catch((e) => {
             return Rx.Observable.of(queryError(e));
@@ -237,19 +232,18 @@ const wfsQueryEpic = (action$, store) =>
         .switchMap(action => {
             return Rx.Observable.merge(
                 getWFSFeature(action.searchUrl, action.filterObj)
+                    .let(interceptOGCError)
                     .switchMap((response) => {
-                        // try to guess if it was a missing id error and try to search again with forced sortOptions
-                        const error = getWFSResponseException(response, 'NoApplicableCode');
-                        if (error) {
-                            return retryWithForcedSortOptions(action, store);
-                        }
                         const state = store.getState();
                         const data = workaroundGEOS7233(response.data, action.filterObj.pagination, paginationInfo.totalFeatures(state));
                         return Rx.Observable.of(querySearchResponse(data, action.searchUrl, action.filterObj));
                     })
                     .startWith(featureLoading(true))
-                    .catch((e) => {
-                        return Rx.Observable.of(queryError(e));
+                    .catch((error) => {
+                        if (error.name === "OGCError" && error.code === 'NoApplicableCode') {
+                            return retryWithForcedSortOptions(action, store);
+                        }
+                        return Rx.Observable.of(queryError(error));
                     })
                     .concat(Rx.Observable.of(featureLoading(false)))
             );
